network for research tree

+++ src/main/java/com/teammoeg/frostedheart/research/Research.java
@@ -246,7 +246,7 @@ public class Research extends FHRegisteredItem implements Writeable {
 	}
 
 	@OnlyIn(Dist.CLIENT)
-	public int getCurrentPoints() {
+	public long getCurrentPoints() {
 		return getData().getTotalCommitted();
 	}
